# _tmp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SakuraFilter API",
    "description": "工业/汽车滤清器产品管理平台 API\r\n\r\n## 模块说明\r\n- **认证**: JWT 登录/刷新/登出/用户管理\r\n- **公开搜索**: 前台产品搜索/详情 (无需认证)\r\n- **后台产品管理**: CRUD (需 admin 角色)\r\n- **字典管理**: 8 类字典 CRUD (需认证)\r\n- **ETL**: 数据导入/状态/进度 (X-Admin-Token)\r\n- **运维**: 健康检查/指标/性能告警\r\n\r\n## 认证方式\r\n1. **JWT Bearer** (推荐): 登录 /api/auth/login 获取 token, 放入 Authorization: Bearer {token}\r\n2. **X-Admin-Token** (ETL/CI 备用): 放入 X-Admin-Token: {token}",
    "contact": {
      "name": "SakuraFilter Team",
      "email": "admin@sakurafilter.dev"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/dead-letter": {
      "get": {
        "tags": [
          "AdminDeadLetter"
        ],
        "summary": "死信队列分页查询 (keyset cursor, 支持 operation/since/recovery_count 过滤)",
        "operationId": "GetDeadLetter",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "operation",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_recovery_count",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "max_recovery_count",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/dead-letter/{id}/recover": {
      "post": {
        "tags": [
          "AdminDeadLetter"
        ],
        "summary": "死信单条恢复 (移回 pending + advisory lock 串行化)",
        "operationId": "RecoverDeadLetter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          }
        }
      }
    },
    "/api/admin/dead-letter/recover-batch": {
      "post": {
        "tags": [
          "AdminDeadLetter"
        ],
        "operationId": "RecoverDeadLetterBatch",
        "parameters": [
          {
            "name": "operation",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastErrorContains",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxRecoveryCount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "X-Admin-Token": {
        "type": "apiKey",
        "description": "Day 8.4: dev 静态 token, 从 appsettings.json 的 Auth:DevStaticToken 读",
        "name": "X-Admin-Token",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer token. 格式: Bearer {token} (登录 /api/auth/login 获取)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "X-Admin-Token": [ ]
    },
    {
      "Bearer": [ ]
    }
  ],
...
}